feat(mapa-camas): filtro por obra social
