+++ src/app/apps/rup/mapa-camas/views/integridad/filtros-inconsistencia/filtros-inconsistencias.component.ts
@@ -1,29 +1,9 @@
-import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
-import { MapaCamasService } from '../../../services/mapa-camas.service';
+import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map, tap } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 import { IntegridadService } from '../integridad.service';
+import { arrayToSet } from '@andes/shared';
 
-function arrayToSet(array, key, itemFn) {
-    const listado = [];
-    array.forEach(elem => {
-        const item = itemFn(elem);
-        if (Array.isArray(item)) {
-            item.forEach(inside => {
-                const index = listado.findIndex(i => i[key] === inside[key]);
-                if (index < 0) {
-                    listado.push(inside);
-                }
-            });
-        } else {
-            const index = listado.findIndex(i => i[key] === item[key]);
-            if (index < 0) {
-                listado.push(item);
-            }
-        }
-    });
-    return listado;
-}
 
 @Component({
     selector: 'app-filtros-inconsistencias',
